### ❓ Type of change

- [ ] 🐞 Bug fix (a non-breaking change that fixes an issue)
- [x] 👌 Enhancement (improving an existing functionality like performance)
- [x] ✨ New feature (a non-breaking change that adds functionality)
- [ ] ⚠️ Breaking change (fix or feature that would cause existing functionality to change)


### 📚 Description

Related to : [How AdonisJS matches routes](https://docs.adonisjs.com/guides/basics/routing#how-adonisjs-matches-routes)

In the following example we **NEVER** going to reach the `/user` route.

```ts
import router from '@adonisjs/core/services/router'

router.get('/:id', ({ response }) => {
  return response.json({ from: 'callback - /:id' })
})

router.get('/user', ({ response }) => {
  return response.json({ from: 'USER ROUTE' })
})
```

After this PR you don't need to worry about the order of your routes. (I hope so, hehe)
> a url that matches the route pattern will have the highest priority


@thetutlage check out [how-adonisjs-matches-routes](https://docs.adonisjs.com/guides/basics/routing#how-adonisjs-matches-routes), I don't think we need this section anymore.
> I can open a PR for the docs if you want.