@RomainLanz 2024-06-07
Hey @TiBianMod! 👋🏻 

Thanks for this proposition.

I see an issue with the requested change, you are adding a loop in a hot-path of the application. Depending on the number of route you have, your code will slow the matching of the wanted route.

I run the benchmark of this repository on my machine with the current code and with your PR, and with the small amount of route in that benchmark, we already see a difference.

**Before**
```
Running 40s test @ http://localhost:4000
100 connections with 10 pipelining factor


┌─────────┬──────┬──────┬───────┬───────┬────────┬─────────┬────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg    │ Stdev   │ Max    │
├─────────┼──────┼──────┼───────┼───────┼────────┼─────────┼────────┤
│ Latency │ 4 ms │ 5 ms │ 10 ms │ 10 ms │ 6.1 ms │ 2.39 ms │ 113 ms │
└─────────┴──────┴──────┴───────┴───────┴────────┴─────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬──────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev    │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┤
│ Req/Sec   │ 143,231 │ 143,231 │ 151,935 │ 155,263 │ 151,472 │ 2,732.47 │ 143,111 │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┤
│ Bytes/Sec │ 26.8 MB │ 26.8 MB │ 28.4 MB │ 29 MB   │ 28.3 MB │ 511 kB   │ 26.8 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴──────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 40

6060k requests in 40s, 1.13 GB read
```

**After**
```
Running 40s test @ http://localhost:4000
100 connections with 10 pipelining factor


┌─────────┬──────┬──────┬───────┬───────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼───────┼─────────┼─────────┼───────┤
│ Latency │ 4 ms │ 5 ms │ 10 ms │ 10 ms │ 6.27 ms │ 2.46 ms │ 47 ms │
└─────────┴──────┴──────┴───────┴───────┴─────────┴─────────┴───────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬───────────┬──────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg       │ Stdev    │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼───────────┼──────────┼─────────┤
│ Req/Sec   │ 138,623 │ 138,623 │ 147,839 │ 151,167 │ 147,478.4 │ 2,990.58 │ 138,574 │
├───────────┼─────────┼─────────┼─────────┼─────────┼───────────┼──────────┼─────────┤
│ Bytes/Sec │ 25.9 MB │ 25.9 MB │ 27.7 MB │ 28.3 MB │ 27.6 MB   │ 560 kB   │ 25.9 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴───────────┴──────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 40

5900k requests in 40s, 1.1 GB read
```

Therefore, I am not really in favor of the change, but maybe @Julien-R44 or @thetutlage will not mind the perf hit.