@TiBianMod 2024-06-07
one question here

we want this route to match `/user` and `/USER` ??

```ts
router.get('/user', ({ response }) => {
  return response.json({ from: 'USER ROUTE' })
})
```