@TiBianMod 2024-06-07
I made a new commit. Let's keep the `sensitive case` of the url for some other PR.

@RomainLanz can you please benchmark it again.

I think now is faster than before.