@TiBianMod 2024-06-07
**Before**

```console
bombardier -c 200 -d 1m http://127.0.0.1:4000
Bombarding http://127.0.0.1:4000 for 1m0s using 200 connection(s)
[==========================================================================] 1m0s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     43758.88    1649.56   47291.27
  Latency        4.57ms     1.63ms   435.38ms
  HTTP codes:
    1xx - 0, 2xx - 2625601, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    10.39MB/s
```

**After**

```console
bombardier -c 200 -d 1m http://127.0.0.1:4000
Bombarding http://127.0.0.1:4000 for 1m0s using 200 connection(s)
[==========================================================================] 1m0s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     45595.91    2224.33   50462.60
  Latency        4.39ms     2.53ms   579.62ms
  HTTP codes:
    1xx - 0, 2xx - 2735811, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    10.83MB/s
```