@thetutlage 2024-06-12
All the major routers be it from Laravel, Rails, or Express have the same top-to-bottom routers matching behavior, therefore it is wise to stick to widely accepted resolution logic.